Move localelist to _app
