+++ frontend/pages/_app.tsx
@@ -3,6 +3,8 @@ import type { AppProps } from 'next/app';
 import Layout from '../components/Layout';
 import { appWithTranslation } from 'next-i18next';
 
+export const localeList = ['common', 'category', 'footer', 'home'];
+
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <Layout>
