--- frontend/pages/localeList.ts
@@ -1,4 +1,0 @@
-/**
- * list of locals to use in this app.
- */
-export const localeList = ['common', 'category', 'footer'];
